+++ WebMvcPluginTour/WebMvcPluginTour/Controllers/TourInfoController.cs
@@ -401,7 +401,7 @@ namespace WebMvcPluginTour.Controllers
                         || !CoreHelper.GetParameter(out var jsonName, body, "Name",
                             JTokenType.String, ref responseModel)
                         || !CoreHelper.GetParameter(out var jsonPrice, body, "Price",
-                            JTokenType.Integer, ref responseModel)
+                            JTokenType.Float, ref responseModel)
                         || !CoreHelper.GetParameter(out var jsonServices, body, "Services",
                             JTokenType.Array, ref responseModel, isNullable: true))
                     {
@@ -423,7 +423,7 @@ namespace WebMvcPluginTour.Controllers
                     _ = int.TryParse(jsonMaxMember?.ToString(), out var maxMember);
                     _ = int.TryParse(jsonTotalDay?.ToString(), out var totalDay);
                     _ = int.TryParse(jsonTotalNight?.ToString(), out var totalNight);
-                    _ = int.TryParse(jsonPrice?.ToString(), out var price);
+                    _ = float.TryParse(jsonPrice?.ToString(), out var price);
                     var serviceIds = jsonServices != null
                         ? JsonConvert.DeserializeObject<string[]>(jsonServices.ToString())
                         : null;
@@ -496,7 +496,7 @@ namespace WebMvcPluginTour.Controllers
                         || !CoreHelper.GetParameter(out var jsonTotalNight, body, "TotalNight",
                             JTokenType.Integer, ref responseModel, isNullable: true)
                         || !CoreHelper.GetParameter(out var jsonPrice, body, "Price",
-                            JTokenType.Integer, ref responseModel, isNullable: true)
+                            JTokenType.Float, ref responseModel, isNullable: true)
                         || !CoreHelper.GetParameter(out var jsonName, body, "Name",
                             JTokenType.String, ref responseModel, isNullable: true)
                         || !CoreHelper.GetParameter(out var jsonTimelines, body, "Timelines",
@@ -528,7 +528,7 @@ namespace WebMvcPluginTour.Controllers
                     var isMaxMemberParse = int.TryParse(jsonMaxMember?.ToString(), out var maxMember);
                     var isTotalDayParsed = int.TryParse(jsonTotalDay?.ToString(), out var totalDay);
                     var isTotalNightParsed = int.TryParse(jsonTotalNight?.ToString(), out var totalNight);
-                    var isPriceParsed = int.TryParse(jsonPrice?.ToString(), out var price);
+                    var isPriceParsed = float.TryParse(jsonPrice?.ToString(), out var price);
                     var services = jsonServices != null
                         ? JsonConvert.DeserializeObject<string[]>(jsonServices.ToString())
                         : null;
@@ -548,7 +548,7 @@ namespace WebMvcPluginTour.Controllers
                         totalDay: isTotalDayParsed ? totalDay : (int?) null,
                         totalNight: isTotalNightParsed ? totalNight: (int?) null,
                         maxMember: isMaxMemberParse ? maxMember : (int?) null,
-                        price: isPriceParsed ? price : (int?) null,
+                        price: isPriceParsed ? price : (float?) null,
                         services: services
                     );
